Add properties from XMLInputFactory.properties file to xmlFactory
When a XMLInputFactory.newInstance() is created the properties are assigned to default values. To override that we can configure them in XMLInputFactory.properties file. Therefore, in the DBUtil.class after creating a new instance using XMLInputFactory.newInstance() the properies should be set by reading the values from XMLInputFactory.properties file.

This is related to wso2/product-ei#1222
Then the user can override the com.ctc.wstx.minTextSegment property (default is 64)