Fix DataServiceCappDeployerServiceComponent not activating issue
bind method is calling before calling the activate method of a
service component in osgi.But in the
DataServiceCappDeployerServiceComponent ComponentContext is set
by the activate method and ComponentContext is used by the
setDataServiceCappDeployer bind method. Null pointer exception
is thrown when setDataServiceCappDeployer is accessing
ComponentContext. Issue is fixed in this commit.

Fixes; wso2/product-ei#3171