Update ODataAdapter.java
INT16 are being converted to Byte instead of Short. Since the max value of Byte is 127, this results in an out of range error.
Eg Columns in SQL Server database that are smallint causes 'Value out of range' error if they have any value that is greater that 127 and thus no OData is returned.